requirements.txt
python-dotenv==1.1.1
django==5.2.4
pillow<=11.3
django-tables2==2.7.5
django-crispy-forms==2.4
crispy-daisyui==0.7.0
celery==5.5.3
pyexiftool==0.5.6
djangorestframework==3.16.1